+++ esp-radio/src/wifi/ap.rs
@@ -6,16 +6,8 @@ use core::fmt;
 use enumset::EnumSet;
 use procmacros::BuilderLite;
 
-use super::{AuthMethod, AuthMethodExt as _, Country, Protocol, SecondaryChannel};
-use crate::{
-    WifiError,
-    sys::include::{
-        wifi_ap_record_t,
-        wifi_second_chan_t_WIFI_SECOND_CHAN_ABOVE,
-        wifi_second_chan_t_WIFI_SECOND_CHAN_BELOW,
-        wifi_second_chan_t_WIFI_SECOND_CHAN_NONE,
-    },
-};
+use super::{AuthMethod, Country, Protocol, SecondaryChannel};
+use crate::{WifiError, sys::include::wifi_ap_record_t};
 
 /// Information about a detected Wi-Fi access point.
 #[derive(Debug, Default, Clone, PartialEq, Eq)]
@@ -163,12 +155,7 @@ pub(crate) fn convert_ap_info(record: &wifi_ap_record_t) -> AccessPointInfo {
         ssid,
         bssid: record.bssid,
         channel: record.primary,
-        secondary_channel: match record.second {
-            wifi_second_chan_t_WIFI_SECOND_CHAN_NONE => SecondaryChannel::None,
-            wifi_second_chan_t_WIFI_SECOND_CHAN_ABOVE => SecondaryChannel::Above,
-            wifi_second_chan_t_WIFI_SECOND_CHAN_BELOW => SecondaryChannel::Below,
-            _ => panic!(),
-        },
+        secondary_channel: SecondaryChannel::from_raw(record.second),
         signal_strength: record.rssi,
         auth_method: Some(AuthMethod::from_raw(record.authmode)),
         country: Country::try_from_c(&record.country),